jmap_core.c: split Blob/* functions into jmap_blob.c
